Proyecto2/docker/Loader/app/requirements.txt
azure-storage-blob==12.16.0
Flask==2.0.1
pymongo== 4.3.3
